refactor(maker): Extract blocked peers functionality into module
